feat: Adds nice!view functionality to repository
Makes the required changes to allow the use of a nice!view. Adding the
nice_view_adapter and nice_view shields to the build.yaml file lead to
errors during the compilation of the files. In order to address the
errors, The Kconfig.defconfig file was edited to remove the ZMK_DISPLAY and
LVGL sections and in the temper.dtsi file, the zephyr,display = &oled
line had to be deleted. I also removed the deep sleep config and replaced it with the display config in the temper.config file